feat(scores): Complete Phase 9 - Atomic Swap to new analytics architecture
Phase 9 Complete: Replaced old analytics implementation with new refactored architecture

**What Changed:**
- Moved 15 reusable chart components to `/score-analytics/components/charts/`
- Updated all imports across score-analytics and page files
- Deleted old bloated files: SingleScoreAnalytics (297 lines), TwoScoreAnalytics (698 lines)
- Swapped analytics pages: analytics-v2.tsx → analytics.tsx
- Removed ANALYTICS_V2 tab from navigation
- Backed up old implementation as analytics-old-backup.tsx

**Files Moved (15 total):**
- ScoreDistribution* (4 files: Chart, Numeric, Boolean, Categorical)
- ScoreTimeSeries* (4 files: Chart, Numeric, Boolean, Categorical)
- Heatmap* (4 files: Heatmap, Cell, Legend, Placeholder)
- MetricCard, ScoreCombobox, ObjectTypeFilter

**Files Deleted (6 total):**
- SingleScoreAnalytics.tsx, TwoScoreAnalytics.tsx
- ComparisonStatistics.tsx, MatchedOnlyToggle.tsx, HeatmapCard.tsx
- analytics/index.ts

**Architecture Impact:**
New structure consolidates all analytics code under `/score-analytics/`:
```
/score-analytics/
  /components/
    /cards/          # Smart cards (consume context)
    /charts/         # Reusable chart components (presentation)
    ScoreAnalyticsHeader.tsx
    ScoreAnalyticsDashboard.tsx
    ScoreAnalyticsProvider.tsx
  /hooks/
    useScoreAnalyticsQuery.ts
  /transformers/
    scoreAnalyticsTransformers.ts
```

**Validation:**
✅ TypeScript: No errors in scores code
✅ Linter: No warnings or errors
✅ All imports updated correctly

**Progress:** Phase 9/10 complete (95% total)
