+++ web/src/features/scores/components/score-analytics/components/charts/HeatmapCell.tsx
