+++ web/src/features/scores/components/score-analytics/components/charts/ScoreCombobox.tsx
